[REF] messages: update ticket system and apply timeout to actions
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@62160 b456876b-0849-0410-b77d-98878d47e9d5

+++ messu-read_sent.php
@@ -14,10 +14,24 @@ $messulib = TikiLib::lib('message');
 $access->check_user($user);
 $access->check_feature('feature_messages');
 $access->check_permission('tiki_p_messages');
-if (isset($_REQUEST["delete"])) {
-	check_ticket('messu-read_sent');
-	$messulib->delete_message($user, $_REQUEST['msgdel'], 'sent');
+$access->checkAuthenticity();
+
+if (!isset($_REQUEST['msgId']) || $_REQUEST['msgId'] == 0) {
+	$smarty->assign('legend', tra("No more messages"));
+	$smarty->assign('mid', 'messu-read_sent.tpl');
+	$smarty->display("tiki.tpl");
+	die;
 }
+
+if ($access->ticketMatch()) {
+	if (isset($_REQUEST['action'])) {
+		$messulib->flag_message($user, $_REQUEST['msgId'], $_REQUEST['action'], $_REQUEST['actionval'], 'sent');
+	}
+	if (isset($_REQUEST["delete"])) {
+		$messulib->delete_message($user, $_REQUEST['msgdel'], 'sent');
+	}
+}
+
 $smarty->assign('sort_mode', $_REQUEST['sort_mode']);
 $smarty->assign('find', $_REQUEST['find']);
 $smarty->assign('flag', $_REQUEST['flag']);
@@ -25,15 +39,7 @@ $smarty->assign('offset', $_REQUEST['offset']);
 $smarty->assign('flagval', $_REQUEST['flagval']);
 $smarty->assign('priority', $_REQUEST['priority']);
 $smarty->assign('legend', '');
-if (!isset($_REQUEST['msgId']) || $_REQUEST['msgId'] == 0) {
-	$smarty->assign('legend', tra("No more messages"));
-	$smarty->assign('mid', 'messu-read_sent.tpl');
-	$smarty->display("tiki.tpl");
-	die;
-}
-if (isset($_REQUEST['action'])) {
-	$messulib->flag_message($user, $_REQUEST['msgId'], $_REQUEST['action'], $_REQUEST['actionval'], 'sent');
-}
+
 // Using the sort_mode, flag, flagval and find get the next and prev messages
 $smarty->assign('msgId', $_REQUEST['msgId']);
 $next = $messulib->get_next_message($user, $_REQUEST['msgId'], $_REQUEST['sort_mode'], $_REQUEST['find'], $_REQUEST['flag'], $_REQUEST['flagval'], $_REQUEST['priority'], 'sent');
@@ -43,7 +49,6 @@ $smarty->assign('prev', $prev);
 // Get the message and assign its data to template vars
 $msg = $messulib->get_message($user, $_REQUEST['msgId'], 'sent');
 $smarty->assign('msg', $msg);
-ask_ticket('messu-read_sent');
 include_once ('tiki-section_options.php');
 include_once ('tiki-mytiki_shared.php');
 $smarty->assign('mid', 'messu-read_sent.tpl');
